• added required escapes for the Tcl string sub-pattern (lossy conversion?) • added \unnnn, \", and \\ as valid escape sequences in strings • added support for $variable and [embedded code] in strings
I don't know Tcl though, so this was based on a quick Google, after Yuhei Kuratomi wrote me about not supporting escaped quotes in strings.

If the matching of escaped characters is exhaustive, it may be nice to add a catch-all for invalid escapes -- we have that in C, and I'm loving it! :)


git-svn-id: http://svn.textmate.org/trunk/Bundles/Tcl.tmbundle@1591 dfb7d73b-c2ec-0310-8fea-fb051d288c6d